feat(oma-pager): extract pager from oma-console into a new crate
Move the terminal pager (Pager, OmaPager, key bindings, highlight, and
theme traits) out of oma-console into a dedicated oma-pager crate.
oma-console now only keeps the print/writer/terminal modules.

The pager is now decoupled from termbg via the PagerTheme trait, and
uses oma-logger instead of spdlog-rs directly.
