+++ oma-console/Cargo.toml
@@ -10,18 +10,9 @@ license = "MIT"
 [dependencies]
 console = { workspace = true, optional = true }
 indicatif = { workspace = true, optional = true }
-spdlog-rs = { workspace = true }
-ratatui = { workspace = true, optional = true }
-# https://github.com/async-rs/async-std/issues/1055
-# https://github.com/tokio-rs/tokio/issues/5535
-termbg = { workspace = true , optional = true }
-ansi-to-tui = { workspace = true, optional = true }
 textwrap = { workspace = true, optional = true }
-aho-corasick = { workspace = true, optional = true }
-tui-input = { workspace = true, optional = true }
 
 [features]
 print = ["dep:textwrap", "dep:console"]
-pager = ["dep:ratatui", "dep:ansi-to-tui", "dep:console", "dep:aho-corasick", "dep:tui-input", "dep:termbg"]
 progress_bar_style = ["dep:indicatif"]
-default = ["print", "pager", "progress_bar_style"]
+default = ["print", "progress_bar_style"]
