+++ src/core/commit_changes.rs
@@ -5,9 +5,10 @@ use bon::Builder;
 use dialoguer::{Confirm, theme::ColorfulTheme};
 use flume::unbounded;
 use jiff::Timestamp;
-use oma_console::{indicatif::HumanBytes, pager::PagerExit};
+use oma_console::indicatif::HumanBytes;
 use oma_history::{DATABASE_PATH, HistoryInfo};
 use oma_logger::{debug, error, info, warn};
+use oma_pager::PagerExit;
 use oma_pm::{
     CommitConfig,
     apt::{InstallEntry, InstallProgressOpt, OmaApt, OmaAptArgs, OmaAptError, RemoveEntry},
