+++ src/tui/mod.rs
@@ -3,8 +3,8 @@ use std::time::Duration;
 use clap::Args;
 use oma_apt_pkg::AptConfig;
 use oma_apt_pkg::search::{IndiciumSearch, OmaSearch, SearchResult, SearchType};
-use oma_console::pager::{exit_tui, prepare_create_tui};
 use oma_logger::{debug, info};
+use oma_pager::{exit_tui, prepare_create_tui};
 use oma_pm::apt::{OmaApt, OmaAptArgs, Upgrade};
 use render::{Task, Tui as TuiInner};
 use zbus::Connection;
